requirements.txt
streamlit
openai
langchain-community
chromadb
pypdf
PyPDF2
faiss-cpu
tiktoken
streamlit-chat
